+++ tailwind.config.js
@@ -1,8 +1,12 @@
-const { colors } = require('tailwindcss/defaultTheme')
+const { colors, screens } = require('tailwindcss/defaultTheme')
 
 module.exports = {
   theme: {
     extend: {
+        screens: {
+            ...screens,
+            'xxl': '1920px'
+        },
         colors: {
             primary: '#09364f',
             secondary: '#f5ba1f',