+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [unreleased]
+
+### Added
+
+- new breakpoint `xxl` to tailwind
+
 ## [v0.1.0] - 15.03.2020
 
 ### Added